hw_7/requirements.txt
pillow==11.2.1
pdf2image==1.17.0